+++ src/Cdn/Cdn/ChangeLog.md
@@ -19,6 +19,7 @@
 -->
 
 ## Upcoming Release
+* Added breaking change messages for all cmdlets in Az.CDN module
 
 ## Version 1.8.2
 * Added breaking change messages for upcoming breaking change release of version 2.0.0